Move fallback attribute to GooglePlayServicesConfiguration
